refactor: split conversation filter internals
Move ConversationFilter builder and runtime responsibilities into focused internal modules while keeping the public class and import path stable.
