+++ polylogue/lib/filter_types.py
@@ -1,0 +1,7 @@
+from __future__ import annotations
+
+from typing import Literal
+
+SortField = Literal["date", "tokens", "messages", "words", "longest", "random"]
+
+__all__ = ["SortField"]
